+++ CHANGES.rst
@@ -30,7 +30,11 @@ New Features
 - ``astropy.io.votable``
 
 - ``astropy.modeling``
+
   - Added ``SmoothlyBrokenPowerLaw1D`` model. [#5656]
+
+  - Add ``n_submodels`` property to compound models, which allows users to get
+    the number of components of a given compound model. [#5747]
 
 - ``astropy.nddata``